refactor: delete queued-run infra dispatcher and remove zero imports
Remove dispatchQueuedRun() from run-service.ts and inline the non-zero
fallback in dispatchQueuedZeroRun() using buildInfraExecutionContext.
This fixes two bugs: non-zero runs no longer incorrectly use
buildZeroExecutionContext and no longer write to zeroRuns table.

Also removes the dead dispatchQueuedRun default parameter from
dispatchCancelSideEffects() and the zeroRuns import from run-service.ts.

Closes #7590
